upload

+++ index.js
@@ -20,15 +20,17 @@ app.use("/api", parseRouter);
 app.use("/api/auth", authRouter);
 
 cron.schedule("59 59 23 * * *", () => {
-	const directory = "./tmp/data";
+	const directory = "/tmp";
 
 	fs.readdir(directory, (err, files) => {
 		if (err) throw err;
 
 		for (const file of files) {
-			fs.unlink(path.join(directory, file), (err) => {
-				if (err) throw err;
-			});
+			if (!file.includes("logs")) {
+				fs.unlink(path.join(directory, file), (err) => {
+					if (err) throw err;
+				});
+			}
 		}
 	});
 });
